+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.1.0
+* Add `threeDSecurePassThru` parameters to `Customer.create`, `PaymentMethod.create`, `CreditCard.create`, `Customer.update`, `PaymentMethod.update` and `CreditCard.update` 
+* Add `recurringCustomerConsent` and `recurringMaxAmount` to `authenticationInsightOptions` for `PaymentMethodNonce.create`
+* Add `DOCUMENT_UPLOAD_FILE_IS_EMPTY` error code
+* phpDocs fixes (thanks @pupitooo)
+
 ## 5.0.0
 * Add `DISPUTE_ACCEPTED`, `DISPUTE_DISPUTED`, and `DISPUTE_EXPIRED` webhook constants
 * Breaking Changes: